+++ extensions/adwords_rate_limiter/ChangeLog.md
@@ -1,5 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [0.1.0] - 2017-02-01
+## [0.1.1] - 2018-01-15
+- Make setting variables public, so clients can set customization via variable names.
+- Use ONE virtual account to handle rate limit of ALL unknown accounts combined.
+- Accepted PR117, Fixing minor typo
+
+## [0.1.0] - 2017-04-03
 - Initial version